fix: copilot feedback
