+++ tests/actor/test_reminder_data.py
@@ -16,8 +16,8 @@ limitations under the License.
 import unittest
 from datetime import timedelta
 
-from dapr.actor.runtime._failure_policy import ActorReminderFailurePolicy
 from dapr.actor.runtime._reminder_data import ActorReminderData
+from dapr.actor.runtime.failure_policy import ActorReminderFailurePolicy
 
 
 class ActorReminderTests(unittest.TestCase):
